main footer design completed
