+++ routes/index.js
@@ -1,4 +1,5 @@
 //const adminRoutes = require('./admin/index');
+const paymentsRoutes = require('./payments');
 const productsRoutes = require('./products');
 const usersRoutes = require('./users');
 
@@ -10,17 +11,17 @@ const mainRoutes = (app) => {
     });
 
     /* customized routes */
-    //app.use("/admin", adminRoutes);         // admin routes
+    app.use("/payment", paymentsRoutes);    // payments routes
     app.use("/product", productsRoutes);    // products routes
     app.use("/user", usersRoutes);          // user routes
 
     /* non existing page configuration */
     app.use("*", (req, res) => {
         res.render('alerts/error', {
-            mainTitle: "404 | ",
+            mainTitle: "Page Not Found •",
             code: 404,
             message: `Page Not Found`,
-            url: req.originalUrl
+            url: req.originalUrl 
         });
     });
 };
